Move grouped dossier information getter to the disposition object.
So it can be reused from the API endpoints as well.
